add day 12 and 13 tasks

+++ Day11-12-13-Tasks/Application/Features/Posts/Handlers/Commands/DeletePostRequestCommandHandler.cs
@@ -1,8 +1,10 @@
 using System.Threading;
 using System.Threading.Tasks;
+using Application.Exceptions;
 using Application.Features.Posts.Requests.Commands;
 using Application.Persistance.Contracts;
 using AutoMapper;
+using Domain;
 using MediatR;
 
 namespace Application.Features.Posts.Handler.Commands
@@ -19,6 +21,10 @@ namespace Application.Features.Posts.Handler.Commands
 
         public async Task<Unit> Handle(DeletePostRequestCommand request, CancellationToken cancellationToken)
         {
+            var post = await _postRepository.GetAsync(request.Id);
+            if (post == null)
+                throw new NotFoundException(nameof(Post), request.Id);
+                
             await _postRepository.DeleteAsync(request.Id);
             return Unit.Value;
         }
